Renamed folders by removing spaces from names.

+++ SuprannuaEngine/SuprannuaEngine.h
@@ -230,7 +230,7 @@ extern double gravityConstant;
 extern bool isGamePaused;
 
 /*Engine time and states*/
-extern int frameCount;
+extern unsigned long frameCount;
 extern double timeCount;
 extern double framesPerSecond;
 extern int passedFrames;
